[RadioButton SwiftUI] Use mocked theme in test.
